Module dbus_serialize::types

Contains the Value and BasicValue enums, as well as traits and helper types for them

Structs

Array
An Array is an ordered sequence of Value objects which must all be of the same variety. That is, it is not value to have a Uint8 and a Uint32 as elements of the same Array.

Dictionary

Path

Signature

Struct
A Struct is an ordered sequence of Value objects, which may be of different varieties

Variant

Enums

BasicValue
BasicValue covers the "basic" D-Bus types, that is those that are allowed to be used as keys in a dictionary.

Value
Root type for any D-Bus value