Struct dbus_serialize::types::Array

pub struct Array {
    pub objects: Vec<Value>,
    // some fields omitted
}

An Array is an ordered sequence of Value objects which must all be of the same variety. That is, it is not value to have a Uint8 and a Uint32 as elements of the same Array.

Fields

objects

Methods

impl Array

fn new(objects: Vec<Value>) -> Array

Create a new array from the given vector of Value. This function may only be used when it is never possible for the input vector to be empty. The reason is that it is impossible to determine the type signature for an empty vector. Use new_with_sig instead.

Panics

If objects.len() is 0, this function will panic.

fn new_with_sig(objects: Vec<Value>, sig: String) -> Array

Create a new array from the given vector of Value. If objects is non-empty, then sig must match the type of the contents.