Struct dbus_serialize::types::Dictionary

pub struct Dictionary {
    pub map: HashMap<BasicValue, Value>,
    // some fields omitted
}

Fields

map

Methods

impl Dictionary

fn new(map: HashMap<BasicValue, Value>) -> Dictionary

Create a new Dictionary from the given map. This function may only be used when it is never possible for the input map to be empty. The reason is that it is impossible to determine the type signature for an empty vector. Use new_with_sig instead.

Panics

If map.len() is 0, this function will panic.

fn new_with_sig(map: HashMap<BasicValue, Value>, sig: String) -> Dictionary

Create a new Dictionary from the given map. If map is non-empty, then sig must match the type of the contents.